% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling_stats.R
\name{binary_model_stats}
\alias{binary_model_stats}
\title{Compute predictive model statistics considering two classes}
\usage{
binary_model_stats(pred, expect, positive = 1)
}
\arguments{
\item{pred}{predicted values}

\item{expect}{expected values}

\item{positive}{value of the positive class}
}
\value{
an array of descriptive statistics
}
\description{
Compute predictive model statistics considering two classes
}
