% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_colors.R
\name{color_darken}
\alias{color_darken}
\title{Darken a vector of colors}
\usage{
color_darken(col, fac = 0.1)
}
\arguments{
\item{col}{a color or a vector of colors (text or hexadecimal).}

\item{fac}{a float number between 0 and 1 that indicate the darkening intensity.}
}
\value{
a vector of darkened color(s).
}
\description{
Darken a vector of colors
}
\examples{
# From a text color
color_darken("red", 0.2)

# From a vector of hexadecimal colors
color_darken(rainbow(10), 0.5)
}
