% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_colors.R
\name{color_lighten}
\alias{color_lighten}
\title{Lighten a vector of colors}
\usage{
color_lighten(col, fac = 0.1)
}
\arguments{
\item{col}{a color or a vector of colors (text or hexadecimal).}

\item{fac}{a float number between 0 and 1 that indicate the lightening intensity.}
}
\value{
a vector of lightened color(s).
}
\description{
Lighten a vector of colors
}
\examples{
# From a text color
color_lighten("red", 0.2)

# From a vector of hexadecimal colors
color_lighten(rainbow(10), 0.5)
}
