% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_colors.R
\name{color_mixer}
\alias{color_mixer}
\title{Mix a vector of colors with another color}
\usage{
color_mixer(col1, col2, p1 = 0.5)
}
\arguments{
\item{col1}{a color or a vector of colors (text or hexadecimal).}

\item{col2}{a unique color to be mixed with col1.}

\item{p1}{proportion of col1, float number between 0 and 1.}
}
\value{
a vector of mixed color(s).
}
\description{
Mix a vector of colors with another color
}
\examples{
# From a text color
color_mixer("red", "blue", 0.2)

# From a vector of hexadecimal colors
color_alpha(rainbow(10), 0.5)
}
