% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_colors.R
\docType{data}
\name{my.tealgreen}
\alias{my.tealgreen}
\title{My teal green}
\format{
An object of class \code{character} of length 1.
}
\usage{
my.tealgreen
}
\description{
My teal green
}
\keyword{datasets}
