% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_colors.R
\name{my_rainbow}
\alias{my_rainbow}
\title{My rainbow color ramp}
\usage{
my_rainbow(n, alpha = 1)
}
\arguments{
\item{n}{number of required colors.}

\item{alpha}{alpha factor.}
}
\value{
a vector of colors.
}
\description{
My rainbow color ramp
}
\examples{
# Get 10 colors
my_rainbow(10)
}
