% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/violin_plot.R
\name{violin_plot}
\alias{violin_plot}
\title{Violin plot}
\usage{
violin_plot(
  x,
  y.lim = NULL,
  y.lab = "",
  mode = "both",
  cols = my_rainbow,
  add.pt = FALSE,
  cex.pt = 0.8,
  add.cnt = FALSE,
  new.plot = TRUE,
  spacer = 0.05,
  smooth.lvl = 1
)
}
\arguments{
\item{x}{a list of values or a data.frame.}

\item{y.lim}{y axis limits.}

\item{y.lab}{y axis label.}

\item{mode}{plotting mode: only \code{left}, only \code{right} or \code{both}.}

\item{cols}{color generator or vector of colors.}

\item{add.pt}{boolean, add points over distributions.}

\item{cex.pt}{points size factor.}

\item{add.cnt}{boolean, add counts.}

\item{new.plot}{boolean, initialize a new plot.}

\item{spacer}{spacer between each distribution.}

\item{smooth.lvl}{smooth level.}
}
\value{
nothing
}
\description{
Performs a violin plot from a list of values.
}
\examples{
# plot normal distribution
dt <- list(
  grp1 = rnorm(100, 0, 1),
  grp2 = rnorm(200, 1, 3)
)
violin_plot(dt)
}
