% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sliding_window.R
\name{sliding_window}
\alias{sliding_window}
\title{Sliding window function}
\usage{
sliding_window(x, sw_len, sw_steps)
}
\arguments{
\item{x}{a vector of values}

\item{sw_len}{length of the window}

\item{sw_steps}{length of the steps between windows}
}
\value{
a vector of values
}
\description{
Sliding window function
}
\examples{
# On a random vector
x <- rnorm(1010)
x.sw <- sliding_window(x, 100, 50)
}
